+++ plugins/module_utils/network/saos10/facts/legacy/base.py
@@ -10,14 +10,30 @@ based on the configuration.
 """
 
 from __future__ import absolute_import, division, print_function
+from ansible.module_utils._text import to_text
 
 __metaclass__ = type
 import platform
-import re
 from ansible_collections.ciena.saos10.plugins.module_utils.network.saos10.saos10 import (
-    run_commands,
+    get_configuration,
     get_capabilities,
+    remove_ns,
 )
+
+from ansible_collections.ansible.netcommon.plugins.module_utils.network.netconf.netconf import (
+    get,
+)
+
+try:
+    from lxml.etree import tostring as xml_to_string
+
+    HAS_LXML = True
+except ImportError:
+    from xml.etree.ElementTree import (
+        tostring as xml_to_string,
+    )
+
+    HAS_LXML = False
 
 
 class FactsBase(object):
@@ -30,29 +46,50 @@ class FactsBase(object):
         self.warnings = list()
         self.responses = None
 
-    def populate(self):
-        self.responses = run_commands(
-            self.module, commands=self.COMMANDS, check_rc=False
-        )
+    def tostring(element, encoding="UTF-8"):
+        if HAS_LXML:
+            return xml_to_string(element, encoding="unicode")
+        else:
+            return to_text(xml_to_string(element, encoding), encoding=encoding)
 
-    def run(self, cmd):
-        return run_commands(self.module, commands=cmd, check_rc=False)
+    def populate(self):
+        raise NotImplementedError
 
 
 class Default(FactsBase):
-
-    COMMANDS = ["show system components"]
-
     def populate(self):
-        super(Default, self).populate()
-        data = self.responses[0]
-        self.facts["serialnum"] = self.parse_serialnum(data)
         self.facts.update(self.platform_facts())
+        config_filter = (
+            '<components xmlns="http://openconfig.net/yang/platform"/>'
+        )
+        reply = get(self.module, filter=("subtree", config_filter))
+        root = remove_ns(reply)
+        serial_number = root.xpath(
+            "/data/components/component[1]/state/serial-no"
+        )[0].text
+        model = root.xpath(
+            "/data/components/component[1]/component-properties/component-property[name = 'hw-model']/value"
+        )[0].text
+        platform = root.xpath("/data/components/component[1]/state/name")[
+            0
+        ].text
+        self.facts["serialnum"] = serial_number
+        self.facts["model"] = model
+        self.facts["platform"] = platform
 
-    def parse_serialnum(self, data):
-        match = re.search(r"\| serial-no +\| +(\S+)", data)
-        if match:
-            return match.group(1)
+        config_filter = '<software-state xmlns="http://www.ciena.com/ns/yang/ciena-software-mgmt"/>'
+        reply = get(self.module, filter=("subtree", config_filter))
+        root = remove_ns(reply)
+        network_os_version = root.xpath(
+            "/data/software-state/running-package/package-version"
+        )[0].text
+        self.facts["version"] = network_os_version
+
+        config_filter = '<system xmlns="http://openconfig.net/yang/system"><config><hostname/></config></system>'
+        reply = get(self.module, filter=("subtree", config_filter))
+        root = remove_ns(reply)
+        hostname = root.xpath("/data/system/config/hostname")[0].text
+        self.facts["hostname"] = hostname
 
     def platform_facts(self):
         platform_facts = {}
@@ -62,7 +99,7 @@ class Default(FactsBase):
 
         platform_facts["system"] = device_info["network_os"]
 
-        for item in ("model", "image", "version", "platform", "hostname"):
+        for item in ("image", "version", "hostname"):
             val = device_info.get("network_os_%s" % item)
             if val:
                 platform_facts[item] = val
@@ -74,9 +111,18 @@ class Default(FactsBase):
 
 
 class Config(FactsBase):
-
-    COMMANDS = ["show running"]
-
     def populate(self):
-        super(Config, self).populate()
-        self.facts["config"] = self.responses[0]
+        config_format = self.module.params["config_format"]
+        config_format = "xml"
+        reply = get_configuration(self.module, format="xml")
+
+        if config_format == "xml":
+            config = xml_to_string(remove_ns(reply))
+
+        elif config_format == "text":
+            raise Exception("text Not yet Implemented")
+
+        elif config_format == "json":
+            raise Exception("json Not yet Implemented")
+
+        self.facts["config"] = config